"...enterprise software can be effective when selected by people who don't have to use it" ?
I think you meant "can't" here?  Also fixed a couple more little typos.

Your analysis of inter-departmental power dynamics rings true, from my vantage point within a company where IT screws over departments all the time. The crap that users have to wade through on a daily basis is truly horrible, especially in departments with less ability to push back -- which in my experience are often so-called "support" staff, vs "line workers", and often majority women.

The bit about disruption is less compelling to me. If in the case of universities the answer is to make them irrelevant, why isn't it in the case of enterprise software, to make capitalist enterprises irrelevant?   ;)